requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
basemap==1.1.0
cycler==0.10.0
decorator==4.3.2
enum34==1.1.6
futures==3.2.0
ipykernel==4.10.0
ipython==4.0.0
ipython-genutils==0.2.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.5
pandas==0.19.2
pathlib2==2.3.3
patsy==0.4.1
pexpect==4.6.0
pickleshare==0.7.5
ptyprocess==0.6.0
pyparsing==2.3.1
pyproj==2.0.2
pyshp==2.1.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==18.0.1
scandir==1.10.0
scipy==0.18.1
seaborn==0.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
statsmodels==0.9.0
subprocess32==3.5.3
tornado==5.1.1
traitlets==4.3.2
